+++ packages/agent/src/polling/index.ts
@@ -5,6 +5,7 @@ import { RequestId } from '../request_id';
 import { toHex } from '../utils/buffer';
 
 export * as strategy from './strategy';
+import { defaultStrategy } from './strategy';
 export { defaultStrategy } from './strategy';
 export type PollStrategy = (
   canisterId: Principal,
@@ -27,7 +28,7 @@ export async function pollForResponse(
   agent: Agent,
   canisterId: Principal,
   requestId: RequestId,
-  strategy: PollStrategy,
+  strategy: PollStrategy = defaultStrategy(),
   // eslint-disable-next-line
   request?: any,
   blsVerify?: CreateCertificateOptions['blsVerify'],